database filter
Re-Architecturing backend:  front-end is sending get-requests to the
back-end, that is querying database for the recent fresh record younger
than 10 minutes or equal. Back-end calls the weather API every 10
minutes to refresh the weather record in the database. So all clients
are talking to the database only without touching weather API, thus we
prevent the situation of overload calls for weather API not to  exceed
the limit of calls per minute.